requirements.txt
opencv-python
opencv-contrib-python
numpy
tqdm
torch
torchvision
Pillow
matplotlib